requirements.txt
asgiref==3.6.0
Django==3.2.10
django-filter==23.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
Markdown==3.4.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
